Make some fix for node2vec and run bash

+++ model/node2vec.py
@@ -114,7 +114,7 @@ def alias_setup(probs):
 	'''
 	K = len(probs)
 	q = np.zeros(K)
-	J = np.zeros(K, dtype=np.int)
+	J = np.zeros(K, dtype=np.int32)
 
 	smaller = []
 	larger = []
